[EI-183] [api] rename: live indicator port 폴더 위치 변경
