+++ api/src/numerical-guidance/application/port/external/krx/load-live-indicator.port.ts
@@ -1,4 +1,4 @@
-import { LiveIndicatorDto } from '../../query/get-live-indicator/live-indicator.dto';
+import { LiveIndicatorDto } from '../../../query/get-live-indicator/live-indicator.dto';
 
 export interface LoadLiveIndicatorPort {
   loadLiveIndicator(indicatorId: string, ticker: string, interval: string, market: string): Promise<LiveIndicatorDto>;
